+++ src/Stack/Path.hs
@@ -51,6 +51,7 @@ path keys =
        global <- GhcPkg.getGlobalDB menv =<< getWhichCompiler
        snaproot <- installationRootDeps
        localroot <- installationRootLocal
+       toolsDir <- bindirCompilerTools
        distDir <- distRelativeDir
        hpcDir <- hpcReportDir
        compiler <- getCompilerPath =<< getWhichCompiler
@@ -84,6 +85,7 @@ path keys =
                                global
                                snaproot
                                localroot
+                               toolsDir
                                distDir
                                hpcDir
                                extra
@@ -108,6 +110,7 @@ data PathInfo = PathInfo
     , piGlobalDb     :: Path Abs Dir
     , piSnapRoot     :: Path Abs Dir
     , piLocalRoot    :: Path Abs Dir
+    , piToolsDir     :: Path Abs Dir
     , piDistDir      :: Path Rel Dir
     , piHpcDir       :: Path Abs Dir
     , piExtraDbs     :: [Path Abs Dir]
@@ -146,6 +149,9 @@ paths =
     , ( "Directory containing the compiler binary (e.g. ghc)"
       , "compiler-bin"
       , T.pack . toFilePathNoTrailingSep . parent . piCompiler )
+    , ( "Directory containing binaries specific to a particular compiler (e.g. intero)"
+      , "compiler-tools-bin"
+      , T.pack . toFilePathNoTrailingSep . piToolsDir )
     , ( "Local bin dir where stack installs executables (e.g. ~/.local/bin)"
       , "local-bin"
       , T.pack . toFilePathNoTrailingSep . configLocalBin . bcConfig . piBuildConfig )